Extract HIR boundary tests
Move the inline HIR lowering boundary regression module out of hir.rs into tests/hir_unit.rs while preserving private module access through a path module declaration.

Lower the compiler source monolith ratchet ceilings to capture the 660-line reduction in hir.rs.

Refs #1254.
